requirements.txt
tensorflow==2.1.0
waymo-open-dataset-tf-2-1-0==1.2.0
torch==1.5.1
filterpy
numpy
